+++ Arcadia/Enums/ArcadiaGameType.swift
@@ -17,6 +17,7 @@ import ArcadiaAtari7800Core
 import ArcadiaNeoGeoPocketCore
 import ArcadiaMasterSystemCore
 import ArcadiaGenesisCore
+import ArcadiaPokemonMiniCore
 import SwiftUI
 
 enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTypeProtocol {
@@ -31,6 +32,7 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
     case gameBoyGame = "GameBoy (Color)"
     case nesGame = "NES"
     case neoGeoPocketGame = "Neo Geo Pocket (Color)"
+    case pokemonMiniGame = "Pokémon Mini"
     case gameGearGame = "Sega Game Gear"
     case genesisGame = "Sega Genesis"
     case masterSystemGame = "Sega Master System - SG 1000"
@@ -59,6 +61,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return "Sega Game Gear"
         case .genesisGame:
             return "Sega Genesis"
+        case .pokemonMiniGame:
+            return "Pokémon Mini"
         }
     }
     
@@ -84,6 +88,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return ["CAST(30 AS INT)"]
         case .genesisGame:
             return ["CAST(33 AS INT)"]
+        case .pokemonMiniGame:
+            return ["CAST(0 AS INT)"]
         }
     }
     
@@ -109,6 +115,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return [UTType(importedAs: "com.davideandreoli.Arcadia.gamefile.gg"), UTType(importedAs: "com.davideandreoli.Arcadia.gamefile.bin")]
         case .genesisGame:
             return [UTType(importedAs: "com.davideandreoli.Arcadia.gamefile.gen"), UTType(importedAs: "com.davideandreoli.Arcadia.gamefile.md"), UTType(importedAs: "com.davideandreoli.Arcadia.gamefile.bin")]
+        case .pokemonMiniGame:
+            return [UTType(importedAs: "com.davideandreoli.Arcadia.gamefile.min")]
         }
         
     }
@@ -135,6 +143,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return ArcadiaMasterSystem()
         case .genesisGame:
             return ArcadiaGenesis()
+        case .pokemonMiniGame:
+            return ArcadiaPokemonMini()
         }
     }
     
@@ -160,6 +170,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return "srm"
         case .genesisGame:
             return "srm"
+        case .pokemonMiniGame:
+            return "eep"
         }
     }
     
@@ -206,6 +218,10 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return [
                 .memorySaveRam : "srm"
             ]
+        case .pokemonMiniGame:
+            return [
+                .memorySaveRam : "eep"
+            ]
         }
         
     }
@@ -232,6 +248,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return Image("game_gear_icon")
         case .genesisGame:
             return Image("sega_genesis_icon")
+        case .pokemonMiniGame:
+            return Image("pokemon_mini_icon")
         }
     }
     
@@ -257,6 +275,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return [.dPad, .twoActionButtons, .start]
         case .genesisGame:
             return [.dPad, .threeActionButtons, .start]
+        case .pokemonMiniGame:
+            return [.dPad, .twoActionButtons, .start, .select, .backButtonsFirstRow]
         }
     }
         
@@ -282,6 +302,8 @@ enum ArcadiaGameType: String, Codable, CaseIterable, Identifiable, ArcadiaGameTy
             return try! AttributedString(markdown: "[Gearsystem](https://github.com/drhelius/Gearsystem)")
         case .genesisGame:
             return try! AttributedString(markdown: "[picodrive](https://github.com/libretro/picodrive)")
+        case .pokemonMiniGame:
+            return try! AttributedString(markdown: "[PokeMini](https://github.com/libretro/PokeMini)")
         }
     }